Feat (ui-config): TailwindCSS 커스텀 변수(xs, scrollbar, slider, progress) 추가
